requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cheroot==6.5.5
CherryPy==18.1.1
connect-sdk==16.0
cryptography==2.6.1
deprecation==2.0.6
et-xmlfile==1.0.1
idna==2.8
jaraco.functools==2.0
jdcal==1.4.1
jwt==0.6.1
marshmallow==2.18.0
more-itertools==7.0.0
openpyxl==2.5.14
packaging==19.0
portend==2.4
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.0
pytz==2019.1
requests==2.21.0
six==1.12.0
tempora==1.14.1
urllib3==1.24.3
zc.lockfile==1.4